backend/requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
django-rest-framework==0.1.0
django-versatileimagefield==2.0
djangorestframework==3.12.4
Pillow==8.2.0
psycopg2==2.8.6
python-magic-bin==0.4.14
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
